v1.1.0-dev1/salad/requirements.txt
typing==3.5.3
avro-cwl==1.8.4 ; python_version>="3"
future ; python_version>="3"
avro==1.8.1 ; python_version<"3"
ruamel.yaml>=0.12.4, <0.15
rdflib==4.2.2
rdflib-jsonld==0.4.0
mistune==0.7.3
CacheControl==0.11.7
lockfile==0.12.2